+++ app/src/test/java/reikai/presentation/library/novels/NovelLibraryItemTest.kt
@@ -27,6 +27,7 @@ class NovelLibraryItemTest {
         sourceBadge = false,
         sourceSite = null,
         sourceIconUrl = null,
+        sourceName = "Novel Arrow",
     )
 
     @Test